refactor(console): split the signup marketing panel into its own file
Move the left-hand testimonial / proof panel (and its code-showcase
widget) out of FirebaseSignup.tsx into SignupMarketingPanel.tsx, so the
signup file is just the form.
